+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project follows the [Semantic Versioning standard](https://semver.org/).
 
 ### Changed
 - Renamed `Frequenzy` to `Frequency`
+- Add limit to `all()`, prevent infinite loops.
 
 ### Deprecated